+++ xenon-common/src/main/java/com/vmware/xenon/services/common/QueryTask.java
@@ -164,7 +164,12 @@ public class QueryTask extends ServiceDocument {
             /**
              * Groups results using the {@link QuerySpecification::groupByTerms}
              */
-            GROUP_BY
+            GROUP_BY,
+
+            /**
+             * Query will return latest versions of documents before {@link QuerySpecification#timeSnapshotBoundaryMicros}
+             */
+            TIME_SNAPSHOT
         }
 
         public enum SortOrder {
@@ -250,6 +255,12 @@ public class QueryTask extends ServiceDocument {
          */
         public transient QueryRuntimeContext context = new QueryRuntimeContext();
 
+        /**
+         * Used with {@link QueryOption#TIME_SNAPSHOT}
+         */
+        @Since(ReleaseConstants.RELEASE_VERSION_1_3_6)
+        public Long timeSnapshotBoundaryMicros;
+
         public static String buildCompositeFieldName(String... fieldNames) {
             StringBuilder sb = new StringBuilder();
             for (String s : fieldNames) {
@@ -332,6 +343,7 @@ public class QueryTask extends ServiceDocument {
             clonedSpec.sortTerm = this.sortTerm;
             clonedSpec.groupSortTerm = this.groupSortTerm;
             clonedSpec.groupSortOrder = this.groupSortOrder;
+            clonedSpec.timeSnapshotBoundaryMicros = this.timeSnapshotBoundaryMicros;
         }
     }
 
@@ -1149,6 +1161,14 @@ public class QueryTask extends ServiceDocument {
          */
         public Builder setIndexLink(String indexLink) {
             this.queryTask.indexLink = indexLink;
+            return this;
+        }
+
+        /**
+         *
+         */
+        public Builder setQueryTimeStamp(Long documentsUpdatedBeforeInMicros) {
+            this.querySpec.timeSnapshotBoundaryMicros = documentsUpdatedBeforeInMicros;
             return this;
         }